Handle function literals and nested functions.
If a function contains a nested function literal, we need to associate
each return statement with the correct function.  A naked return
statement in the function literal may be okay if it has no named return
values.

To track state, we create a new visitor each time we enter a function or
function literal.